fix: Resolve remaining CI build failures across all platforms
macOS (AppleClang):
- Add clang diagnostic pragmas for -Wnested-anon-types and
  -Wgnu-anonymous-struct in cpu_context.h, input.h, instance_state.h
- Change HeadlessBackend move ops from = default to = delete
  (members contain non-movable std::mutex/std::atomic)
- Fix tautological SIZE_MAX comparison in SAFE_MULTIPLY_OR_ERROR
  macro by casting to std::size_t

Linux (Clang 18):
- Add libstdc++-13-dev to CI package installs for C++23 support
- Add missing #include <chrono> in POSIX IPC source files

Windows (MSVC):
- Increase build timeout-minutes from 15/20 to 30

+++ engine/include/aibox/safe_arithmetic.h
@@ -51,7 +51,7 @@ safe_multiply_3(std::size_t a, std::size_t b, std::size_t c) noexcept {
  */
 #define SAFE_MULTIPLY_OR_ERROR(a, b, result_var) \
     do { \
-        if ((b) != 0 && (a) > SIZE_MAX / (b)) { \
+        if ((b) != 0 && static_cast<std::size_t>(a) > SIZE_MAX / (b)) { \
             DOSBOXX_ERROR(DOSBOXX_ERR_INVALID_STATE, \
                 "Integer overflow: " #a " * " #b); \
         } \
